requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML/AI
torch>=2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
coremltools==7.0

# Graph & Vector DB
py2neo==2021.2.4
faiss-cpu==1.7.4
networkx==2.8.8
node2vec==0.4.6

# Data processing
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Privacy
cryptography==41.0.7

# Streaming & messaging
kafka-python==2.0.2
redis==5.0.1

# Database
psycopg2-binary==2.9.9
influxdb-client==1.38.0

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# API
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.1
gym==0.26.2
